pkp/pkp-lib#6502 Fix fatal errors when accessing file that doesn't exist
This commit checks that a file exists before trying to retrieve details
like the mimetype or file size. To improve support for third-party file
storage systems, it reduces the number of cases where the file needs
to be accessed on the file system to return information about it. These
changes include:

- Mimetype is now stored in the files table.
- The size property is no longer returned in API requests.
- File mimetypes are now migrated from the old database when upgrading
	to 3.3.
- The docSize property was removed from usageEvent data.

+++ classes/tombstone/DataObjectTombstoneDAO.inc.php
@@ -57,8 +57,8 @@ class DataObjectTombstoneDAO extends DAO {
 			'SELECT * FROM data_object_tombstones WHERE data_object_id = ?',
 			[(int) $dataObjectId]
 		);
-
-		return $this->_fromRow((array) $result->current());
+		$row = $result->current();
+		return $row ? $this->_fromRow((array) $row) : null;
 	}
 
 	/**
@@ -111,11 +111,12 @@ class DataObjectTombstoneDAO extends DAO {
 	/**
 	 * Delete DataObjectTombstone by data object id.
 	 * @param $dataObjectId int
-	 * @return boolean
 	 */
 	function deleteByDataObjectId($dataObjectId) {
 		$dataObjectTombstone = $this->getByDataObjectId($dataObjectId);
-		return $this->deleteById($dataObjectTombstone->getId());
+		if ($dataObjectTombstone) {
+			$this->deleteById($dataObjectTombstone->getId());
+		}
 	}
 
 	/**
